requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
branca==0.6.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.3
django-allauth==0.55.0
django-autoslug==1.9.9
django-bootstrap-icons==0.8.3
django-bootstrap-v5==1.0.11
django-ckeditor==6.6.1
django-js-asset==2.1.0
folium==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4

django-crispy-forms~=2.1